修改日志中协程id的获取

+++ base/util.h
@@ -11,7 +11,7 @@ namespace BASE
 {
 
 pid_t getThreadId();    //获取线程id
-u_int32_t getFiberId(); //获取携程id
+uint64_t getFiberId(); //获取携程id
 
 //获取堆栈信息放到v中 size是获取多少层 skip要跳过多少层
 void Backtrace(std::vector<std::string> & v, int size, int skip);
